requirements.txt
alembic==1.0.10
aniso8601==6.0.0
APScheduler==3.6.1
atomicwrites==1.3.0
attrs==19.1.0
blinker==1.4
CacheControl==0.12.5
certifi==2019.6.16
chardet==3.0.4
Click==7.0
deprecation==2.0.7
entrypoints==0.3
flake8==3.7.8
Flask==1.0.3
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.2
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpickle==0.9.6
jsonschema==3.0.1
Mako==1.0.11
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.3
mccabe==0.6.1
more-itertools==8.1.0
msgpack==0.6.1
packaging==19.1
paho-mqtt==1.4.0
passlib==1.7.1
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.2
pyrsistent==0.15.2
pytest==5.1.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
redis==3.2.1
requests==2.22.0
rq==1.1.0
sightengine==1.4.0
six==1.12.0
SQLAlchemy==1.3.4
tzlocal==2.0.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.6.0
